+++ src/web/assets/commerceui/src/js/components/OrderBlock.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="order-block">
+    <div class="order-block py">
         <slot></slot>
     </div>
 </template>
@@ -7,7 +7,6 @@
 <style lang="scss">
     .order-block {
         border-top: 1px solid #eee;
-        padding: 20px 0;
 
         h3 {
             font-weight: bold;